Modularized Control Schemes
Created Controller folder to hold all Controller files.

Controller now stores references to different control schemes, and keeps track of which is the active scheme.

Control schemes implement the event handlers declared in the interface ControlScheme which is defined in Controller.js.

Much of the old functionality of Controller was moved into the first ControlScheme called BuildingScheme.

Main now exports the canvas dimensions.